util: remove file only of O_CREAT was specified

+++ util.cc
@@ -100,7 +100,9 @@ bool writeBufToFile(const char* filename, const void* buf, size_t len, int open_
 	if (!writeToFd(fd, buf, len)) {
 		PLOG_E("Couldn't write '%zu' bytes to file '%s' (fd='%d')", len, filename, fd);
 		close(fd);
-		unlink(filename);
+		if (open_flags & O_CREAT) {
+			unlink(filename);
+		}
 		return false;
 	}
